Add stats to the impact report page
