+++ tbx/core/blocks.py
@@ -743,6 +743,45 @@ class EmbedBlock(WagtailEmbedBlock):
         template = "patterns/molecules/streamfield/blocks/embed_block.html"
 
 
+class NumericStatisticsBlock(blocks.StructBlock):
+    headline_number = blocks.CharBlock(
+        max_length=255,
+        help_text=mark_safe("A numerical value e.g. <strong>30%</strong>"),
+    )
+    description = blocks.TextBlock(
+        help_text=mark_safe(
+            "Text to describe the change e.g. <strong>Reduction in accessibility errors</strong>"
+        )
+    )
+    further_details = blocks.TextBlock(
+        required=False,
+        help_text=mark_safe(
+            "Text to give more information, e.g. <strong>Over 80% of pages</strong>"
+        ),
+    )
+
+    class Meta:
+        icon = "table"
+
+
+class TextualStatisticsBlock(blocks.StructBlock):
+    headline_text = blocks.CharBlock(
+        max_length=255,
+        help_text=mark_safe(
+            "Describe a general improvement, e.g. <strong>Reduction in accessibility issues</strong>"
+        ),
+    )
+    further_details = blocks.TextBlock(
+        required=False,
+        help_text=mark_safe(
+            "Text to give more information, e.g. <strong>Over 80% of pages</strong>"
+        ),
+    )
+
+    class Meta:
+        icon = "info-circle"
+
+
 class TypedTableBlock(blocks.StructBlock):
     table = WagtailTypedTableBlock(
         [
